enh(engine): state_history and get_category moved from notifier
state_history is moved to checkable
get_category is moved to notification_manager

The notification_manager singleton now has an explicitly controlled lifetime
(init()/deinit(), mirroring checks::checker). A Meyers singleton was destroyed
before the global host/service objects at program exit, so the notifier
destructors calling notification_manager::forget() touched a destroyed
absl::flat_hash_map (abseil "Use of destroyed hash table" abort on engine
shutdown). forget() is now a guarded static, like checker::forget().

+++ engine/inc/com/centreon/engine/retention/applier/utils.hh
@@ -30,7 +30,7 @@ namespace utils {
 bool is_command_exist(std::string const& command_line);
 void set_state_history(
     std::vector<int> const& values,
-    std::array<int, MAX_STATE_HISTORY_ENTRIES>& state_history);
+    std::array<uint16_t, MAX_STATE_HISTORY_ENTRIES>& state_history);
 }  // namespace utils
 }  // namespace applier
 }  // namespace retention
